move slot falllback data to media step
